% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_nn.R
\name{.get_neighbors}
\alias{.get_neighbors}
\title{Core Logic for Randomized Neighbor Extraction in the Presence of Repetition (which occurs with high probability for discrete data)}
\usage{
.get_neighbors(i, dists_mat, idxs_mat, n)
}
\arguments{
\item{i}{Integer, the index of the i-th observation in X}

\item{dists_mat}{A matrix of distances between all X_i and X_j}

\item{idxs_mat}{A matrix of all neighbor indices sorted by their distances for each observation}

\item{n}{Integer, the number of observations in the random variable X}
}
\value{
A list including a vector of first two groups of neighbors for X_i and a vector of the first two neighbors of X_i
}
\description{
This function finds the indices of the first two nearest neighbors for the i-th row in matrix X.
To handle duplicate distances (e.g., from repeated points), it groups neighbors based on their distance:
  - First group: all neighbors with the smallest non-zero distance (excluding the point itself if the distance is zero)
  - Second group: all neighbors with the next smallest distance
Both groups are randomly shuffled before selecting the first two overall neighbors.
}
\keyword{internal}
