% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ford.R
\name{ford_main}
\alias{ford_main}
\title{Variable selection by the ford algorithm
ford_main is the core function for ford, implementing the
variable selection algorithm based on the Integrated R square dependence coefficient \code{\link{irdec}}.}
\usage{
ford_main(
  Y,
  X,
  dist.type.X = "continuous",
  num_features = NULL,
  stop = TRUE,
  numCores = parallel::detectCores()
)
}
\arguments{
\item{Y}{Vector of responses (length n)}

\item{X}{Matrix of predictors (n by p)}

\item{dist.type.X}{A string specifying the distribution type of X: either "continuous" or "discrete". Default is "continuous".}

\item{num_features}{Number of variables to be selected, cannot be larger than p. The default value is NULL and in that
case it will be set equal to p. If stop == TRUE (see below), then num_features is irrelevant.}

\item{stop}{Stops at the first instance of negative irdec, if TRUE.}

\item{numCores}{number of cores that are going to be used for parallelizing the process.}
}
\value{
An object of class "ford", with three attributes a vector of selected covariates,
their names and their cumulative dependence value with Y in decreasing order of predictive power.
}
\description{
Variable selection by the ford algorithm
ford_main is the core function for ford, implementing the
variable selection algorithm based on the Integrated R square dependence coefficient \code{\link{irdec}}.
}
\details{
ford is a forward stepwise algorithm that uses the Integrated R square dependence coefficient (\code{\link{irdec}})
at each step, instead of the multiple correlation coefficient
as in ordinary forward stepwise. If stop == TRUE, the process is stopped at the first instance of
nonpositive irdec, thereby selecting subset of variables. Otherwise, a set of covariates of size
num_features, ordered according to predictive power (as measured by irdec) is produced.
}
\references{
Azadkia, M. and Roudaki, P. (2025). A NEW MEASURE OF DEPENDENCE: INTEGRATED R2
\url{https://arxiv.org/pdf/???}.
}
\seealso{
\code{\link{irdec}}
}
\author{
Mona Azadkia, Pouya Roudaki
}
