% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irdec.R
\name{irdec}
\alias{irdec}
\title{Estimate the Integrated R-squared Dependence Coefficient (IRDEC)}
\usage{
irdec(Y, X, dist.type.X = "continuous", na.rm = TRUE)
}
\arguments{
\item{Y}{A vector of length n.}

\item{X}{A vector or matrix of length n (or with n rows).}

\item{dist.type.X}{A string specifying the distribution type of X: either "continuous" or "discrete". Default is "continuous".}

\item{na.rm}{Logical; if TRUE, missing values (NAs) will be removed. Default is TRUE.}
}
\value{
The Integrated R-squared Dependence Coefficient (IRDEC) between Y and X.
}
\description{
The Integrated R-squared Dependence Coefficient (IRDEC) is a measure of dependence between
a random variable Y and a random vector X, based on an i.i.d. sample of (Y, X).
The estimated coefficient is asymptotically guaranteed to lie between 0 and 1.
The coefficient is asymmetrical; that is, irdec(X, Y) ≠ irdec(Y, X).
The coefficient equals 0 if and only if X is independent of Y, and it equals 1 if and only if
Y is a measurable function of X.
This coefficient has several applications; for example, it can be used for variable selection, as demonstrated in the \code{\link{ford}} function.
}
\details{
The value returned by `irdec` can be positive or negative for finite samples,
but asymptotically, it is guaranteed to be between 0 and 1.
A small value indicates low conditional dependence between Y and X, while a high value indicates strong conditional dependence.
The `irdec` function is used by the \code{\link{foci}} function for variable selection.
}
\examples{
n = 1000
x <- matrix(runif(n * 3), nrow = n)
y <- (x[, 1] + x[, 2]) \%\% 1
irdec(y, x[, 1])
irdec(y, x[, 2])
irdec(y, x[, 3])
}
\references{
Azadkia, M. and Roudaki, P. (2025). A NEW MEASURE OF DEPENDENCE: INTEGRATED R2
\url{https://arxiv.org/pdf/???}.
}
\seealso{
\code{\link{ford}}, \code{\link{foci}}, \code{\link[XICOR]{xicor}}
}
\author{
Mona Azadkia, Pouya Roudaki
}
